Use base image from donor booted ostree

+++ ostree-backup.sh
@@ -1,8 +1,12 @@
 #!/bin/bash
 
-backup_repo=${BACKUP_REPO:-quay.io/jpolo/ostmagic}
+backup_repo=${1:-$BACKUP_REPO}
 backup_tag=backup
+base_tag=base
+parent_tag=parent
 backup_refspec=$backup_repo:$backup_tag
+base_refspec=$backup_repo:$base_tag
+parent_refspec=$backup_repo:$parent_tag
 
 log_it(){
     echo $@ | tr [:print:] -
@@ -10,11 +14,20 @@ log_it(){
     echo $@ | tr [:print:] -
 }
 
+if [[ -z "$backup_repo" ]]; then
+    echo "ERROR. Backup repo is empty"
+    exit 1
+fi
+
 sudo crictl ps -o json| jq -r '.containers[] | .imageRef' > /var/tmp/containers.list
 
-log_it Stopping kubelet and crio
+log_it Stopping kubelet
 systemctl stop kubelet
-crictl ps -q | xargs crictl stop
+log_it Stopping containers
+crictl ps -q | xargs --no-run-if-empty crictl stop --timeout 5
+log_it Waiting for containers to stop
+while crictl ps -q | grep -q .; do sleep 1; done
+log_it Stopping crio
 systemctl stop crio
 
 log_it Cleaning /var
@@ -35,13 +48,26 @@ tar czf /var/tmp/backup/var.tgz \
     --exclude='/var/lib/cni/bin/*' \
     --selinux \
     /var
-ostree admin config-diff | awk '{print "/etc/" $2}' | xargs tar czf /var/tmp/backup/etc.tgz
+ostree admin config-diff | awk '{print "/etc/" $2}' | xargs tar czf /var/tmp/backup/etc.tgz --selinux
 rpm-ostree status > /var/tmp/backup/rpm-ostree.status
-ostree commit --branch backup /var/tmp/backup
+rpm-ostree status -v --json > /var/tmp/backup/rpm-ostree.json
+cp /etc/machine-config-daemon/currentconfig /var/tmp/backup/mco-currentconfig.json
+ostree commit --branch $backup_tag /var/tmp/backup
 
 # Create credentials for pushing the generated image
 mkdir /root/.docker
 cp backup-secret.json /root/.docker/config.json
 
 log_it Encapsulating and pushing backup OCI
-ostree container encapsulate backup registry:$backup_refspec --repo /ostree/repo --label ostree.bootable=true
+ostree container encapsulate $backup_tag registry:$backup_refspec --repo /ostree/repo --label ostree.bootable=true
+
+log_it Encapsulating and pushing base OCI
+base_commit=$(rpm-ostree status -v --json | jq -r '.deployments[] | select(.booted == true).checksum')
+ostree container encapsulate $base_commit registry:$base_refspec --repo /ostree/repo --label ostree.bootable=true
+
+# If there's a parent to that commit, also encapsulate it
+if [[ "$(rpm-ostree status -v --json | jq -r '.deployments[] | select(.booted == true)| has("base-checksum")')" == "true" ]]; then
+    log_it Parent commit found for base, encapsulating and pushing OCI
+    parent_commit=$(rpm-ostree status -v --json | jq -r '.deployments[] | select(.booted == true)."base-checksum"')
+    ostree container encapsulate $parent_commit registry:$parent_refspec --repo /ostree/repo --label ostree.bootable=true
+fi
